+++ src/plugin/src/identifier-case/local-plan.js
@@ -5,7 +5,7 @@ import {
     toPosixPath,
     resolveProjectPathInfo
 } from "../../../shared/path-utils.js";
-import { createMetricsTracker } from "../../../shared/reporting.js";
+import { createMetricsTracker } from "../reporting/index.js";
 import {
     isNonEmptyString,
     getNonEmptyString,
